change to 32bit sine

+++ include/synth/oscillator.h
@@ -5,6 +5,6 @@
 
 void osc_init(void);
 void osc_set_frequency(uint32_t freq_hz);
-int16_t osc_next_sample(void);
+int32_t osc_next_sample(void);
 
 #endif
