+++ include/synth/synth.h
@@ -5,6 +5,6 @@
 
 void synth_init(void);
 void synth_set_note(int note_on, uint32_t freq_hz);
-int16_t synth_next_sample(void);
+int32_t synth_next_sample(void);
 
 #endif
